+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Version 1.x.y - 2021-MM-DD
 - Add cost estimation to all adapters
 - Add Datasette adapter
 - Remove ``csv://`` and ``datasette+`` prefixes to simply URIs
+- Add ``has_table`` method to dialects
 
 Version 1.0.0 - 2021-08-18
 ==========================